24665 - mudado o POM por causa das alteraçoes no cpf, ainda com o problema do timeout no require (+1 squashed commits)
Squashed commits:

[209d37ef] BACKLOG-24665 - still has the timeout problem and some resources are missing (ccc)

+++ core/impl/src/main/java/pt/webdetails/cdf/dd/structure/DashboardStructure.java
@@ -174,8 +174,9 @@ public class DashboardStructure implements IDashboardStructure {
 
     // Any data sources?
     if ( cdaRenderer.isEmpty() ) {
-      access.deleteFile( cdaFileName );
-
+      if( access.fileExists( cdaFileName ) ) {
+        access.deleteFile(cdaFileName);
+      }
     } else {
       // throws Exception ????
       String cdaText = cdaRenderer.render();
